refactor(desktop): modularize main process
