+++ packages/desktop/src/main/files/apps.ts
