+++ packages/desktop/src/main/files/attachment-picker.test.ts
