+++ packages/desktop/src/main/index.ts
@@ -1,358 +1,94 @@
-import { randomUUID } from "node:crypto"
-import { mkdirSync, rmSync } from "node:fs"
-import * as http from "node:http"
-import { homedir, tmpdir } from "node:os"
-import { join } from "node:path"
-import { getCACertificates, setDefaultCACertificates } from "node:tls"
-import type { Event } from "electron"
-import { app, BrowserWindow } from "electron"
-
+import { app } from "electron"
 import { Deferred, Effect, Fiber } from "effect"
-import contextMenu from "electron-context-menu"
-
 import type { ServerReadyData } from "../shared/ipc-contract"
-import { checkAppExists, resolveAppPath } from "./apps"
-import { CHANNEL, VERSION } from "./constants"
+import { checkAppExists, resolveAppPath } from "./files/apps"
+import { registerIpcHandlers, registerUpdaterIpcHandlers, registerWslIpcHandlers } from "./ipc"
 import {
-  registerIpcHandlers,
-  registerUpdaterIpcHandlers,
-  registerWslIpcHandlers,
-  sendDeepLinks,
-  sendMenuCommand,
-} from "./ipc"
-import { forwardInitializationFailure } from "./initialization"
-import { exportDebugLogs, initCrashReporter, initLogging, startNetLog, write as writeLog } from "./logging"
-import { createMenu } from "./menu"
-import {
-  finishFirstLaunchOnboarding,
-  initializeFirstLaunchOnboarding,
-  isFirstLaunchOnboardingPending,
-} from "./onboarding"
-import { getDefaultServerUrl, preferAppEnv, setDefaultServerUrl } from "./server"
-import { createUpdaterIpc, setupAutoUpdater, showUpdaterDialog } from "./updater"
-import { safeWebContentsURL } from "./window-state"
-import {
-  getLastFocusedWindow,
-  registerRendererProtocol,
-  setRelaunchHandler,
-  setAppQuitting,
-  setBackgroundColor,
-  setDockIcon,
-  restoreMainWindows,
-} from "./windows"
-import { createWslIpc } from "./wsl/ipc"
-import { cleanupStoreFiles } from "./store-cleanup"
-import { startBackgroundCli } from "./background-cli"
-import { setNativeTranslations } from "./native-translations"
-
-const APP_NAMES: Record<string, string> = {
-  dev: "OpenCode Dev",
-  beta: "OpenCode Beta",
-  prod: "OpenCode",
-}
-const APP_IDS: Record<string, string> = {
-  dev: "ai.opencode.desktop.dev",
-  beta: "ai.opencode.desktop.beta",
-  prod: "ai.opencode.desktop",
-}
-const TEST_ONBOARDING = process.env.OPENCODE_TEST_ONBOARDING === "1"
-const jsCallStackFeature = "DocumentPolicyIncludeJSCallStacksInCrashReports"
-
-let logger: ReturnType<typeof initLogging>
-
-const pendingDeepLinks: string[] = []
-
-function useEnvProxy() {
-  try {
-    // Electron 41.2 runs Node 24.14.1; latest @types/node@24 is 24.12.2.
-    ;(http as any).setGlobalProxyFromEnv()
-  } catch (error) {
-    logger.warn("failed to load proxy environment", error)
-  }
-}
-
-function emitDeepLinks(urls: string[]) {
-  if (urls.length === 0) return
-  pendingDeepLinks.push(...urls)
-  const win = getLastFocusedWindow()
-  if (win) sendDeepLinks(win, urls)
-}
-
-function ensureLoopbackNoProxy() {
-  const loopback = ["127.0.0.1", "localhost", "::1"]
-  const upsert = (key: string) => {
-    const items = (process.env[key] ?? "")
-      .split(",")
-      .map((value: string) => value.trim())
-      .filter((value: string) => Boolean(value))
-
-    for (const host of loopback) {
-      if (items.some((value: string) => value.toLowerCase() === host)) continue
-      items.push(host)
-    }
-
-    process.env[key] = items.join(",")
-  }
-
-  upsert("NO_PROXY")
-  upsert("no_proxy")
-}
+  acquireApplicationLock,
+  configureApplication,
+  loadProxyEnvironment,
+  preferApplicationEnvironment,
+  prepareDesktop,
+} from "./lifecycle/environment"
+import { createApplicationLifecycle } from "./lifecycle"
+import { finishFirstLaunchOnboarding, isFirstLaunchOnboardingPending } from "./lifecycle/onboarding"
+import { exportDebugLogs, startNetworkLogging, writeLog } from "./native/logging"
+import { createMenu, sendMenuCommand } from "./native/menu"
+import { setNativeTranslations } from "./native/translations"
+import { startBackgroundCli } from "./service/background-service"
+import { forwardInitializationFailure } from "./service/initialization"
+import { getDefaultServerUrl, setDefaultServerUrl } from "./service/server-settings"
+import { createUpdaterIpc, setupAutoUpdater, showUpdaterDialog, startAutoUpdater } from "./updater"
+import { getLastFocusedWindow, setBackgroundColor } from "./windows"
+import { startWsl } from "./wsl/start"
 
 const main = Effect.gen(function* () {
-  contextMenu({ showSaveImageAs: true, showLookUpSelection: false, showSearchWithGoogle: false })
-
-  // on macOS apps run in `/` which can cause issues with ripgrep
-  try {
-    process.chdir(homedir())
-  } catch {}
-
-  process.env.OPENCODE_DISABLE_EMBEDDED_WEB_UI = "true"
-
-  const appId = app.isPackaged ? APP_IDS[CHANNEL] : "ai.opencode.desktop.dev"
-  const onboardingTestRoot = ((): string | undefined => {
-    if (!TEST_ONBOARDING) return
-
-    const root = join(tmpdir(), `opencode-onboarding-${randomUUID()}`)
-    rmSync(root, { recursive: true, force: true })
-    ;["data", "config", "cache", "state", "desktop", "session"].forEach((dir) =>
-      mkdirSync(join(root, dir), { recursive: true }),
-    )
-    process.env.OPENCODE_DB = ":memory:"
-    process.env.XDG_DATA_HOME = join(root, "data")
-    process.env.XDG_CONFIG_HOME = join(root, "config")
-    process.env.XDG_CACHE_HOME = join(root, "cache")
-    process.env.XDG_STATE_HOME = join(root, "state")
-    return root
-  })()
-  app.setName(app.isPackaged ? APP_NAMES[CHANNEL] : "OpenCode Dev")
-  app.setAppUserModelId(appId)
-  app.setPath(
-    "userData",
-    onboardingTestRoot ? join(onboardingTestRoot, "desktop") : join(app.getPath("appData"), appId),
-  )
-  if (onboardingTestRoot) app.setPath("sessionData", join(onboardingTestRoot, "session"))
-  initializeFirstLaunchOnboarding(app.getPath("userData"))
-  logger = initLogging()
-  initCrashReporter()
-
-  let stopWslServers = async () => {}
-  const relaunch = () => {
-    setAppQuitting()
-    void stopWslServers().finally(() => {
-      app.relaunch()
-      app.quit()
-    })
-  }
-
-  try {
-    setDefaultCACertificates([...new Set([...getCACertificates("default"), ...getCACertificates("system")])])
-  } catch (error) {
-    logger.warn("failed to load system certificates", error)
-  }
-
-  logger.log("app starting", {
-    version: VERSION,
-    packaged: app.isPackaged,
-    onboardingTest: Boolean(onboardingTestRoot),
-  })
-
-  ensureLoopbackNoProxy()
-  useEnvProxy()
-  app.commandLine.appendSwitch("proxy-bypass-list", "<-loopback>")
-  const features = app.commandLine.getSwitchValue("enable-features")
-  app.commandLine.appendSwitch("enable-features", features ? `${jsCallStackFeature},${features}` : jsCallStackFeature)
-  if (!app.isPackaged) app.commandLine.appendSwitch("remote-debugging-port", "9222")
-
-  if (!app.requestSingleInstanceLock()) {
-    app.quit()
-    return
-  }
-
-  preferAppEnv()
-
-  app.on("second-instance", (_event: Event, argv: string[]) => {
-    const urls = argv.filter((arg: string) => arg.startsWith("opencode://"))
-    if (urls.length) {
-      logger.log("deep link received via second-instance", { urls })
-      emitDeepLinks(urls)
-    }
-    const win = getLastFocusedWindow()
-    if (win) {
-      win.show()
-      win.focus()
-    }
-  })
-
-  app.on("open-url", (event: Event, url: string) => {
-    event.preventDefault()
-    logger.log("deep link received via open-url", { url })
-    emitDeepLinks([url])
-  })
-
-  app.on("before-quit", () => {
-    setAppQuitting()
-    void stopWslServers()
-  })
-
-  app.on("will-quit", () => {
-    setAppQuitting()
-    void stopWslServers()
-  })
-
-  app.on("child-process-gone", (_event, details) => {
-    writeLog("utility", "child process gone", { details }, "error")
-  })
-
-  app.on("render-process-gone", (_event, webContents, details) => {
-    writeLog("window", "app render process gone", { url: safeWebContentsURL(webContents), details }, "error")
-  })
-
-  setRelaunchHandler(() => {
-    relaunch()
-  })
-
-  for (const signal of ["SIGINT", "SIGTERM"] as const) {
-    process.on(signal, () => {
-      setAppQuitting()
-      void stopWslServers().finally(() => app.quit())
-    })
-  }
-
+  const logger = configureApplication()
+  if (!acquireApplicationLock()) return
+  preferApplicationEnvironment(logger)
+  const lifecycle = createApplicationLifecycle(logger)
   const serverReady = Deferred.makeUnsafe<ServerReadyData, unknown>()
 
   yield* Effect.promise(() => app.whenReady())
+  yield* prepareDesktop(logger)
 
-  yield* Effect.promise(() => cleanupStoreFiles(app.getPath("userData"))).pipe(
-    Effect.tap((result) =>
-      Effect.sync(() => {
-        if (result.deleted.length === 0) return
-        logger.log("cleaned scoped store files", { count: result.deleted.length, scanned: result.scanned })
-      }),
-    ),
-    Effect.catch((error) =>
-      Effect.sync(() => {
-        logger.warn("failed to clean scoped store files", error)
-      }),
-    ),
-  )
-  app.setAsDefaultProtocolClient("opencode")
-  registerRendererProtocol()
-  setDockIcon()
-  const updater = setupAutoUpdater(() => stopWslServers())
-  const menuDeps = {
+  const updater = setupAutoUpdater(lifecycle.prepareToRestart)
+  const menu = {
     trigger: (id: string) => {
       const win = getLastFocusedWindow()
       if (win) sendMenuCommand(win, id)
     },
     checkForUpdates: () => void showUpdaterDialog(updater),
-    relaunch,
+    relaunch: lifecycle.relaunch,
   }
   registerIpcHandlers({
-    relaunch,
+    relaunch: lifecycle.relaunch,
     awaitInitialization: Effect.fnUntraced(
       function* () {
         logger.log("awaiting server ready")
-        const res = yield* Deferred.await(serverReady)
-        logger.log("server ready", { url: res.url })
-        return res
+        const result = yield* Deferred.await(serverReady)
+        logger.log("server ready", { url: result.url })
+        return result
       },
-      (e) => Effect.runPromise(e),
+      (effect) => Effect.runPromise(effect),
     ),
-    consumeInitialDeepLinks: () => pendingDeepLinks.splice(0),
-    getDefaultServerUrl: () => getDefaultServerUrl(),
-    setDefaultServerUrl: (url) => setDefaultServerUrl(url),
+    consumeInitialDeepLinks: lifecycle.consumeInitialDeepLinks,
+    getDefaultServerUrl,
+    setDefaultServerUrl,
     isFirstLaunchOnboardingPending,
     finishFirstLaunchOnboarding,
-    checkAppExists: (appName) => checkAppExists(appName),
+    checkAppExists,
     resolveAppPath: async (appName) => resolveAppPath(appName),
     showUpdater: () => showUpdaterDialog(updater),
-    setBackgroundColor: (color) => setBackgroundColor(color),
-    exportDebugLogs: () => exportDebugLogs(),
+    setBackgroundColor,
+    exportDebugLogs,
     recordFatalRendererError: (error) => writeLog("renderer", "fatal renderer error", { ...error }, "error"),
     setNativeTranslations: (bundle) => {
-      if (setNativeTranslations(bundle)) createMenu(menuDeps)
+      if (setNativeTranslations(bundle)) createMenu(menu)
     },
   })
   registerUpdaterIpcHandlers(createUpdaterIpc(updater))
-  void updater.start()
-  const updateTimer = setInterval(() => void updater.check(), 10 * 60 * 1000)
-  updateTimer.unref()
-  app.once("will-quit", () => clearInterval(updateTimer))
-  yield* Effect.promise(() => startNetLog()).pipe(
-    Effect.catch((error) =>
-      Effect.sync(() => {
-        logger.warn("failed to start net log", error)
-      }),
-    ),
-  )
+  startAutoUpdater(updater)
+  yield* Effect.promise(() => startNetworkLogging())
 
   const loadingTask = yield* Effect.gen(function* () {
-    ensureLoopbackNoProxy()
-    useEnvProxy()
-
+    loadProxyEnvironment(logger)
     logger.log("starting v2 background service")
     const background = yield* Effect.promise(() => startBackgroundCli(logger))
-    stopWslServers = yield* Effect.promise(() => startWslServers(background))
-
+    const wsl = yield* Effect.promise(() => startWsl(background, logger))
+    registerWslIpcHandlers(wsl.ipc)
+    wsl.start()
+    lifecycle.setWslShutdown(wsl.stop)
     yield* Deferred.succeed(serverReady, {
       url: background.url,
       username: background.username,
       password: background.password,
     })
-
     logger.log("loading task finished")
   }).pipe(forwardInitializationFailure(serverReady), Effect.forkChild)
 
   yield* Fiber.await(loadingTask)
-
-  app.on("window-all-closed", () => {
-    if (process.platform === "darwin") return
-    app.quit()
-  })
-  app.on("activate", () => {
-    if (BrowserWindow.getAllWindows().length > 0) return
-    restoreMainWindows()
-  })
-
-  const windows = restoreMainWindows()
-  if (windows.length) createMenu(menuDeps)
+  if (lifecycle.restoreWindows().length) createMenu(menu)
 })
-
-async function startWslServers(cli: { version: string; wslBuild?: { script: string; output: string } }) {
-  if (process.platform !== "win32") {
-    registerWslIpcHandlers(createWslIpc())
-    return async () => {}
-  }
-
-  const { createWslServersController } = await import("./wsl/servers")
-  const { spawnWslSidecar } = await import("./wsl/sidecar")
-  const local = cli.wslBuild
-  const controller = createWslServersController({
-    cli: { version: cli.version },
-    installCli: local
-      ? async (distro) => {
-          const { buildLocalWslCli } = await import("./wsl/local")
-          const { installWslCli } = await import("./wsl/runtime")
-          await installWslCli(distro, {
-            version: cli.version,
-            binary: await buildLocalWslCli({ ...local, version: cli.version }),
-          })
-        }
-      : undefined,
-    spawnSidecar: async (distro) => {
-      logger.log("spawning wsl sidecar", { distro })
-      return spawnWslSidecar(distro, {
-        onLine: (line) => logger.log("wsl sidecar", { distro, stream: line.stream, text: line.text }),
-      })
-    },
-    logger: {
-      log: (message, meta) => logger.log(message, meta),
-      error: (message, meta) => logger.error(message, meta),
-    },
-  })
-  registerWslIpcHandlers(createWslIpc(controller))
-  controller.startConfiguredServers()
-  return async () => controller.stopServers()
-}
 
 Effect.runFork(main)
