+++ packages/desktop/src/main/ipc.ts
@@ -1,13 +1,9 @@
-import { execFile } from "node:child_process"
-import { stat } from "node:fs/promises"
-import { basename, join } from "node:path"
-import { app, BrowserWindow, clipboard, dialog, ipcMain, shell } from "electron"
+import { BrowserWindow, ipcMain } from "electron"
 import type { IpcMainEvent, IpcMainInvokeEvent } from "electron"
 import { parseDesktopNativeBundle, type DesktopNativeBundle } from "@opencode-ai/app/i18n/desktop-native"
 
 import {
   Ipc,
-  sendIpcEvent,
   type FatalRendererError,
   type IpcInvoke,
   type IpcInvokeArgs,
@@ -15,21 +11,11 @@ import {
   type IpcSend,
   type ServerReadyData,
 } from "../shared/ipc-contract"
-import { runDesktopMenuAction } from "./desktop-menu-actions"
-import { setForceFocus } from "./debug"
-import { assertAttachmentBudget, createPickedFileAuthorizations } from "./attachment-picker"
-import { getStore, removeStoreFileIfEmpty } from "./store"
-import {
-  getPinchZoomEnabled,
-  getWindowID,
-  openExternalURL,
-  openLocalFileURL,
-  setPinchZoomEnabled,
-  setTitlebar,
-  updateTitlebar,
-} from "./windows"
-import { createDesktopDraftStore } from "./draft-store"
-import { nativeT } from "./native-translations"
+import { createFileCapabilities, openExternalURL, openLocalFileURL } from "./files"
+import { setForceFocus } from "./native/debug"
+import { runDesktopMenuAction } from "./native/menu-actions"
+import { createDesktopStorage } from "./storage"
+import { getPinchZoomEnabled, getWindowID, setPinchZoomEnabled, setTitlebar, updateTitlebar } from "./windows"
 import type { UpdaterIpc } from "./updater"
 import type { WslIpc } from "./wsl/ipc"
 
@@ -49,13 +35,6 @@ function on<Channel extends keyof IpcSend>(
   ipcMain.on(channel, listener)
 }
 
-const pickerFilters = (ext?: string[]) => {
-  if (!ext || ext.length === 0) return undefined
-  return [{ name: nativeT("desktop.dialog.files"), extensions: ext }]
-}
-
-const pickedFiles = createPickedFileAuthorizations()
-
 type Deps = {
   relaunch: () => void
   awaitInitialization: () => Promise<ServerReadyData>
@@ -74,10 +53,8 @@ type Deps = {
 }
 
 export function registerIpcHandlers(deps: Deps) {
-  const drafts = createDesktopDraftStore(join(app.getPath("userData"), "drafts.sqlite"))
-  app.on("before-quit", () => drafts.flush())
-  app.once("will-quit", () => drafts.close())
-  app.on("browser-window-created", (_event, win) => win.on("session-end", () => drafts.flush()))
+  const files = createFileCapabilities()
+  const storage = createDesktopStorage()
 
   handle(Ipc.app.awaitInitialization, () => deps.awaitInitialization())
   handle(Ipc.app.consumeInitialDeepLinks, () => deps.consumeInitialDeepLinks())
@@ -103,124 +80,29 @@ export function registerIpcHandlers(deps: Deps) {
     deps.setNativeTranslations(bundle)
   })
   handle(Ipc.storage.get, (_event, name, key) => {
-    try {
-      const store = getStore(name)
-      const value = store.get(key)
-      if (value === undefined || value === null) return null
-      return typeof value === "string" ? value : JSON.stringify(value)
-    } catch {
-      return null
-    }
+    return storage.get(name, key)
   })
-  handle(Ipc.storage.set, (_event, name, key, value) => {
-    getStore(name).set(key, value)
-  })
-  handle(Ipc.storage.delete, (_event, name, key) => {
-    getStore(name).delete(key)
-    void removeStoreFileIfEmpty(name)
-  })
-  handle(Ipc.storage.clear, (_event, name) => {
-    getStore(name).clear()
-    void removeStoreFileIfEmpty(name)
-  })
-  handle(Ipc.storage.keys, (_event, name) => {
-    const store = getStore(name)
-    return Object.keys(store.store)
-  })
-  handle(Ipc.storage.length, (_event, name) => {
-    const store = getStore(name)
-    return Object.keys(store.store).length
-  })
-  handle(Ipc.drafts.get, (_event, key) => drafts.get(key))
-  handle(Ipc.drafts.set, (_event, key, value) => drafts.set(key, value))
-  handle(Ipc.drafts.delete, (_event, key) => drafts.set(key, null))
-  handle(Ipc.drafts.putBlob, (_event, data) => drafts.putBlob(new Uint8Array(data)))
-  handle(Ipc.drafts.getBlob, (_event, id) => {
-    const data = drafts.getBlob(id)
-    return data ? new Uint8Array(data).buffer : null
-  })
+  handle(Ipc.storage.set, (_event, name, key, value) => storage.set(name, key, value))
+  handle(Ipc.storage.delete, (_event, name, key) => storage.deleteValue(name, key))
+  handle(Ipc.storage.clear, (_event, name) => storage.clear(name))
+  handle(Ipc.storage.keys, (_event, name) => storage.keys(name))
+  handle(Ipc.storage.length, (_event, name) => storage.length(name))
+  handle(Ipc.drafts.get, (_event, key) => storage.drafts.get(key))
+  handle(Ipc.drafts.set, (_event, key, value) => storage.drafts.set(key, value))
+  handle(Ipc.drafts.delete, (_event, key) => storage.drafts.set(key, null))
+  handle(Ipc.drafts.putBlob, (_event, data) => storage.drafts.putBlob(data))
+  handle(Ipc.drafts.getBlob, (_event, id) => storage.drafts.getBlob(id))
 
-  handle(Ipc.files.openDirectoryPicker, async (_event, opts) => {
-    const result = await dialog.showOpenDialog({
-      properties: ["openDirectory", ...(opts?.multiple ? ["multiSelections" as const] : []), "createDirectory"],
-      title: opts?.title ?? nativeT("desktop.dialog.chooseFolder"),
-      defaultPath: opts?.defaultPath,
-    })
-    if (result.canceled) return null
-    return opts?.multiple ? result.filePaths : result.filePaths[0]
-  })
-
-  handle(Ipc.files.openFilePicker, async (event, opts) => {
-    const result = await dialog.showOpenDialog({
-      properties: ["openFile", ...(opts?.multiple ? ["multiSelections" as const] : [])],
-      title: opts?.title ?? nativeT("desktop.dialog.chooseFile"),
-      defaultPath: opts?.defaultPath,
-      filters: pickerFilters(opts?.extensions),
-    })
-    if (result.canceled) return null
-    const files = await Promise.all(
-      result.filePaths.map(async (filePath) => ({
-        path: filePath,
-        name: basename(filePath),
-        size: (await stat(filePath)).size,
-      })),
-    )
-    assertAttachmentBudget(files)
-    const token = pickedFiles.add(event.sender.id, result.filePaths)
-    return { token, files }
-  })
-
-  handle(Ipc.files.readPickedFile, async (event, token, filePath) => {
-    return pickedFiles.read(event.sender.id, token, filePath)
-  })
-
-  handle(Ipc.files.releasePickedFiles, (event, token) => {
-    pickedFiles.release(event.sender.id, token)
-  })
-
-  handle(Ipc.files.saveFilePicker, async (_event, opts) => {
-    const result = await dialog.showSaveDialog({
-      title: opts?.title ?? nativeT("desktop.dialog.saveFile"),
-      defaultPath: opts?.defaultPath,
-    })
-    if (result.canceled) return null
-    return result.filePath ?? null
-  })
-
-  on(Ipc.files.openExternal, (_event, url) => {
-    openExternalURL(url)
-  })
-
-  on(Ipc.files.openLocalFile, (_event, url) => {
-    openLocalFileURL(url)
-  })
-
-  handle(Ipc.files.openPath, async (_event, path, app) => {
-    if (!app) return shell.openPath(path)
-    await new Promise<void>((resolve, reject) => {
-      const [cmd, args] =
-        process.platform === "darwin" ? (["open", ["-a", app, path]] as const) : ([app, [path]] as const)
-      execFile(cmd, args, (err) => (err ? reject(err) : resolve()))
-    })
-  })
-
-  handle(Ipc.files.revealPath, async (_event, path) => {
-    const exists = await stat(path).then(
-      () => true,
-      () => false,
-    )
-    if (!exists) return false
-    shell.showItemInFolder(path)
-    return true
-  })
-
-  handle(Ipc.files.readClipboardImage, () => {
-    const image = clipboard.readImage()
-    if (image.isEmpty()) return null
-    const buffer = new Uint8Array(image.toPNG()).buffer
-    const size = image.getSize()
-    return { buffer, width: size.width, height: size.height }
-  })
+  handle(Ipc.files.openDirectoryPicker, (_event, options) => files.openDirectoryPicker(options))
+  handle(Ipc.files.openFilePicker, (event, options) => files.openFilePicker(event.sender.id, options))
+  handle(Ipc.files.readPickedFile, (event, token, path) => files.readPickedFile(event.sender.id, token, path))
+  handle(Ipc.files.releasePickedFiles, (event, token) => files.releasePickedFiles(event.sender.id, token))
+  handle(Ipc.files.saveFilePicker, (_event, options) => files.saveFilePicker(options))
+  on(Ipc.files.openExternal, (_event, url) => openExternalURL(url))
+  on(Ipc.files.openLocalFile, (_event, url) => openLocalFileURL(url))
+  handle(Ipc.files.openPath, (_event, path, app) => files.openPath(path, app))
+  handle(Ipc.files.revealPath, (_event, path) => files.revealPath(path))
+  handle(Ipc.files.readClipboardImage, () => files.readClipboardImage())
 
   handle(Ipc.window.getId, (event) => {
     const win = BrowserWindow.fromWebContents(event.sender)
@@ -299,12 +181,4 @@ export function registerWslIpcHandlers(wsl: WslIpc) {
   handle(Ipc.wsl.addServer, (_event, value) => wsl.addServer(value))
   handle(Ipc.wsl.removeServer, (_event, value) => wsl.removeServer(value))
   handle(Ipc.wsl.startServer, (_event, value) => wsl.startServer(value))
-}
-
-export function sendMenuCommand(win: BrowserWindow, id: string) {
-  sendIpcEvent(win.webContents, Ipc.menu.command, id)
-}
-
-export function sendDeepLinks(win: BrowserWindow, urls: string[]) {
-  sendIpcEvent(win.webContents, Ipc.app.deepLink, urls)
 }
