+++ packages/desktop/src/main/lifecycle/onboarding.ts
@@ -2,10 +2,10 @@ import { existsSync, readdirSync } from "node:fs"
 import { mkdir } from "node:fs/promises"
 import { join } from "node:path"
 import { app } from "electron"
-import { getStore } from "./store"
-import { FIRST_LAUNCH_ONBOARDING_COMPLETE_KEY } from "./store-keys"
-import { write as writeLog } from "./logging"
-import { hasExistingAppState } from "./install-state"
+import { writeLog } from "../native/logging"
+import { hasExistingAppState } from "../storage/install-state"
+import { FIRST_LAUNCH_ONBOARDING_COMPLETE_KEY } from "../storage/keys"
+import { getStore } from "../storage/store"
 
 const DEFAULT_PROJECT_DIR = "Default Project"
