+++ packages/desktop/src/main/native/debug.ts
