+++ packages/desktop/src/main/native/translations.ts
