--- packages/desktop/src/main/server.ts
@@ -1,49 +1,0 @@
-import { getLogger } from "./logging"
-import { getUserShell, loadShellEnv } from "./shell-env"
-import { getStore } from "./store"
-import { DEFAULT_SERVER_URL_KEY } from "./store-keys"
-
-export function getDefaultServerUrl(): string | null {
-  const value = getStore().get(DEFAULT_SERVER_URL_KEY)
-  return typeof value === "string" ? value : null
-}
-
-export function setDefaultServerUrl(url: string | null) {
-  if (url) {
-    getStore().set(DEFAULT_SERVER_URL_KEY, url)
-    return
-  }
-
-  getStore().delete(DEFAULT_SERVER_URL_KEY)
-}
-
-export function preferAppEnv() {
-  const shell = process.platform === "win32" ? null : getUserShell()
-  const shellEnv = shell ? loadShellEnv(shell, getLogger()) : null
-  if (!shellEnv?.XDG_STATE_HOME) delete process.env.XDG_STATE_HOME
-  Object.assign(process.env, {
-    ...shellEnv,
-    OPENCODE_EXPERIMENTAL_ICON_DISCOVERY: "true",
-    OPENCODE_EXPERIMENTAL_FILEWATCHER: "true",
-    OPENCODE_CLIENT: "desktop",
-  })
-}
-
-export async function checkHealth(url: string, password?: string | null): Promise<boolean> {
-  const headers = new Headers()
-  if (password) {
-    const auth = Buffer.from(`opencode:${password}`).toString("base64")
-    headers.set("authorization", `Basic ${auth}`)
-  }
-
-  try {
-    const res = await fetch(new URL("/api/health", url), {
-      method: "GET",
-      headers,
-      signal: AbortSignal.timeout(3000),
-    })
-    return res.ok
-  } catch {
-    return false
-  }
-}
