+++ packages/desktop/src/main/service/cli-version.ts
