+++ packages/desktop/src/main/service/health.ts
@@ -1,0 +1,18 @@
+export async function checkHealth(url: string, password?: string | null): Promise<boolean> {
+  const headers = new Headers()
+  if (password) {
+    const auth = Buffer.from(`opencode:${password}`).toString("base64")
+    headers.set("authorization", `Basic ${auth}`)
+  }
+
+  try {
+    const res = await fetch(new URL("/api/health", url), {
+      method: "GET",
+      headers,
+      signal: AbortSignal.timeout(3000),
+    })
+    return res.ok
+  } catch {
+    return false
+  }
+}
