+++ packages/desktop/src/main/service/initialization.test.ts
