+++ packages/desktop/src/main/service/initialization.ts
