+++ packages/desktop/src/main/service/server-settings.ts
@@ -1,0 +1,16 @@
+import { DEFAULT_SERVER_URL_KEY } from "../storage/keys"
+import { getStore } from "../storage/store"
+
+export function getDefaultServerUrl(): string | null {
+  const value = getStore().get(DEFAULT_SERVER_URL_KEY)
+  return typeof value === "string" ? value : null
+}
+
+export function setDefaultServerUrl(url: string | null) {
+  if (url) {
+    getStore().set(DEFAULT_SERVER_URL_KEY, url)
+    return
+  }
+
+  getStore().delete(DEFAULT_SERVER_URL_KEY)
+}
