+++ packages/desktop/src/main/service/shell-env.test.ts
