+++ packages/desktop/src/main/service/shell-env.ts
