+++ packages/desktop/src/main/storage/cleanup.ts
