+++ packages/desktop/src/main/storage/install-state.test.ts
