+++ packages/desktop/src/main/storage/install-state.ts
