+++ packages/desktop/src/main/windows/registry.ts
