+++ packages/desktop/src/main/windows/state.ts
