+++ packages/desktop/src/main/wsl/runtime.ts
@@ -3,8 +3,8 @@ import { existsSync } from "node:fs"
 import { join } from "node:path"
 import * as pty from "@lydell/node-pty"
 import type { WslDistroProbe, WslInstalledDistro, WslOnlineDistro, WslRuntimeCheck } from "@opencode-ai/app/wsl/types"
-import { parseCliVersion } from "../cli-version"
-import { nativeT } from "../native-translations"
+import { nativeT } from "../native/translations"
+import { parseCliVersion } from "../service/cli-version"
 
 export type WslCommandLine = {
   stream: "stdout" | "stderr"
